+++ src/main/java/com/epam/reportportal/service/launch/AbstractJoinedLaunch.java
@@ -41,8 +41,7 @@ public class AbstractJoinedLaunch extends LaunchImpl {
 			ScheduledExecutorService service) {
 		Random r = new Random();
 		int delay = updateInterval > Integer.MAX_VALUE ? r.nextInt(Integer.MAX_VALUE) : r.nextInt((int) updateInterval);
-		return service.scheduleWithFixedDelay(
-				() -> launchIdLock.updateInstanceUuid(instanceUuid),
+		return service.scheduleWithFixedDelay(() -> launchIdLock.updateInstanceUuid(instanceUuid),
 				delay,
 				updateInterval,
 				TimeUnit.MILLISECONDS